ua/requirements.txt
fastapi
pydantic
pydantic_ai
python-dotenv
uvicorn
redis